Move HttpErrorResponseApi out of folder
